1.8 fix bug of DatumFloat, change its inner value type from float to double
Former-commit-id: ba9a8e7f025a4243d243e6b64546b3df285ad151

+++ src/ra/AggregatorAvg.java
@@ -20,7 +20,7 @@ public class AggregatorAvg extends Aggregator{
 	public Datum getValue(String key) {
 		Datum total = sumer.getValue(key);
 		Datum number = counter.getValue(key);		
-		return new DatumFloat((float)(total.getNumericValue()/number.getNumericValue()));
+		return new DatumFloat(total.getNumericValue()/number.getNumericValue());
 	}
 
 	@Override
